+++ libs/rxjs/utility/src/index.ts
@@ -8,7 +8,7 @@
 /* istanbul ignore file */
 export { debounceWithQuery } from './lib/debounce-with-query';
 export { decodeJWT } from './lib/decode-jwt';
-export { distance } from './lib/distance';
+export { length } from 'libs/rxjs/utility/src/lib/length';
 export { fromReadableStream } from './lib/from-readable-stream';
 export { fromFetchWithProgress } from './lib/from-fetch-with-progress';
 export { hexToRGBA } from './lib/hex-to-rgba';
@@ -25,5 +25,5 @@ export { weight } from './lib/weight';
 
 export * from './types/utility';
 export { Temperatures } from './types/temperature';
-export { Distances } from './types/distance';
+export { Lengths } from 'libs/rxjs/utility/src/types/length';
 export { Weights } from './types/weight';