+++ libs/rxjs/utility/src/lib/length.spec.ts
@@ -1,7 +1,7 @@
 import { marbles } from 'rxjs-marbles/jest';
-import { distance, Distances } from '@rxjs-ninja/rxjs-utility';
+import { length, Lengths } from '@rxjs-ninja/rxjs-utility';
 
-describe('distance', () => {
+describe('length', () => {
   // Centimeters
   describe('cm', () => {
     it(
@@ -18,7 +18,7 @@ describe('distance', () => {
           b: 3.281,
           c: 13.779,
         });
-        m.expect(input.pipe(distance(Distances.CM, Distances.FEET))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.CM, Lengths.FEET))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -36,7 +36,7 @@ describe('distance', () => {
           b: 39.37,
           c: 165.354,
         });
-        m.expect(input.pipe(distance(Distances.CM, Distances.INCHES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.CM, Lengths.INCHES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -55,7 +55,7 @@ describe('distance', () => {
           b: 0.001,
           c: 0.004,
         });
-        m.expect(input.pipe(distance(Distances.CM, Distances.KM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.CM, Lengths.KM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -74,7 +74,7 @@ describe('distance', () => {
           b: 1,
           c: 4.2,
         });
-        m.expect(input.pipe(distance(Distances.CM, Distances.METERS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.CM, Lengths.METERS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -93,7 +93,7 @@ describe('distance', () => {
           b: 0.0006,
           c: 0.0026,
         });
-        m.expect(input.pipe(distance(Distances.CM, Distances.MILES, 4))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.CM, Lengths.MILES, 4))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -112,7 +112,7 @@ describe('distance', () => {
           b: 1.09,
           c: 4.59,
         });
-        m.expect(input.pipe(distance(Distances.CM, Distances.YARDS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.CM, Lengths.YARDS, 2))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -134,7 +134,7 @@ describe('distance', () => {
           b: 3048.037,
           c: 12801.756,
         });
-        m.expect(input.pipe(distance(Distances.FEET, Distances.CM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.FEET, Lengths.CM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -152,7 +152,7 @@ describe('distance', () => {
           b: 1200,
           c: 5040,
         });
-        m.expect(input.pipe(distance(Distances.FEET, Distances.INCHES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.FEET, Lengths.INCHES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -171,7 +171,7 @@ describe('distance', () => {
           b: 0.03,
           c: 0.128,
         });
-        m.expect(input.pipe(distance(Distances.FEET, Distances.KM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.FEET, Lengths.KM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -190,7 +190,7 @@ describe('distance', () => {
           b: 30.48,
           c: 128.018,
         });
-        m.expect(input.pipe(distance(Distances.FEET, Distances.METERS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.FEET, Lengths.METERS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -209,7 +209,7 @@ describe('distance', () => {
           b: 0.0189,
           c: 0.0795,
         });
-        m.expect(input.pipe(distance(Distances.FEET, Distances.MILES, 4))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.FEET, Lengths.MILES, 4))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -228,7 +228,7 @@ describe('distance', () => {
           b: 33.333,
           c: 139.999,
         });
-        m.expect(input.pipe(distance(Distances.FEET, Distances.YARDS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.FEET, Lengths.YARDS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -250,7 +250,7 @@ describe('distance', () => {
           b: 254,
           c: 1066.8,
         });
-        m.expect(input.pipe(distance(Distances.INCHES, Distances.CM, 1))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.INCHES, Lengths.CM, 1))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -268,7 +268,7 @@ describe('distance', () => {
           b: 8.333,
           c: 35,
         });
-        m.expect(input.pipe(distance(Distances.INCHES, Distances.FEET))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.INCHES, Lengths.FEET))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -287,7 +287,7 @@ describe('distance', () => {
           b: 0.003,
           c: 0.011,
         });
-        m.expect(input.pipe(distance(Distances.INCHES, Distances.KM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.INCHES, Lengths.KM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -306,7 +306,7 @@ describe('distance', () => {
           b: 2.54,
           c: 10.668,
         });
-        m.expect(input.pipe(distance(Distances.INCHES, Distances.METERS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.INCHES, Lengths.METERS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -325,7 +325,7 @@ describe('distance', () => {
           b: 0.002,
           c: 0.007,
         });
-        m.expect(input.pipe(distance(Distances.INCHES, Distances.MILES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.INCHES, Lengths.MILES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -344,7 +344,7 @@ describe('distance', () => {
           b: 2.78,
           c: 11.67,
         });
-        m.expect(input.pipe(distance(Distances.INCHES, Distances.YARDS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.INCHES, Lengths.YARDS, 2))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -366,7 +366,7 @@ describe('distance', () => {
           b: 10000000,
           c: 42000000,
         });
-        m.expect(input.pipe(distance(Distances.KM, Distances.CM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.KM, Lengths.CM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -384,7 +384,7 @@ describe('distance', () => {
           b: 328080,
           c: 1377936,
         });
-        m.expect(input.pipe(distance(Distances.KM, Distances.FEET))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.KM, Lengths.FEET))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -403,7 +403,7 @@ describe('distance', () => {
           b: 3937000,
           c: 16535400,
         });
-        m.expect(input.pipe(distance(Distances.KM, Distances.INCHES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.KM, Lengths.INCHES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -422,7 +422,7 @@ describe('distance', () => {
           b: 100000,
           c: 420000,
         });
-        m.expect(input.pipe(distance(Distances.KM, Distances.METERS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.KM, Lengths.METERS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -441,7 +441,7 @@ describe('distance', () => {
           b: 62.137,
           c: 260.975,
         });
-        m.expect(input.pipe(distance(Distances.KM, Distances.MILES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.KM, Lengths.MILES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -460,7 +460,7 @@ describe('distance', () => {
           b: 109360,
           c: 459312,
         });
-        m.expect(input.pipe(distance(Distances.KM, Distances.YARDS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.KM, Lengths.YARDS, 2))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -482,7 +482,7 @@ describe('distance', () => {
           b: 10000,
           c: 42000,
         });
-        m.expect(input.pipe(distance(Distances.METERS, Distances.CM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.METERS, Lengths.CM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -500,7 +500,7 @@ describe('distance', () => {
           b: 328.08,
           c: 1377.936,
         });
-        m.expect(input.pipe(distance(Distances.METERS, Distances.FEET))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.METERS, Lengths.FEET))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -519,7 +519,7 @@ describe('distance', () => {
           b: 3937,
           c: 16535.4,
         });
-        m.expect(input.pipe(distance(Distances.METERS, Distances.INCHES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.METERS, Lengths.INCHES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -538,7 +538,7 @@ describe('distance', () => {
           b: 0.1,
           c: 0.42,
         });
-        m.expect(input.pipe(distance(Distances.METERS, Distances.KM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.METERS, Lengths.KM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -557,7 +557,7 @@ describe('distance', () => {
           b: 0.062,
           c: 0.261,
         });
-        m.expect(input.pipe(distance(Distances.METERS, Distances.MILES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.METERS, Lengths.MILES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -576,7 +576,7 @@ describe('distance', () => {
           b: 109.36,
           c: 459.31,
         });
-        m.expect(input.pipe(distance(Distances.METERS, Distances.YARDS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.METERS, Lengths.YARDS, 2))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -598,7 +598,7 @@ describe('distance', () => {
           b: 16093471,
           c: 67592578,
         });
-        m.expect(input.pipe(distance(Distances.MILES, Distances.CM, 0))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.MILES, Lengths.CM, 0))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -616,7 +616,7 @@ describe('distance', () => {
           b: 528000,
           c: 2217600,
         });
-        m.expect(input.pipe(distance(Distances.MILES, Distances.FEET))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.MILES, Lengths.FEET))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -635,7 +635,7 @@ describe('distance', () => {
           b: 6336000,
           c: 26611200,
         });
-        m.expect(input.pipe(distance(Distances.MILES, Distances.INCHES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.MILES, Lengths.INCHES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -654,7 +654,7 @@ describe('distance', () => {
           b: 160.935,
           c: 675.926,
         });
-        m.expect(input.pipe(distance(Distances.MILES, Distances.KM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.MILES, Lengths.KM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -673,7 +673,7 @@ describe('distance', () => {
           b: 160934.709,
           c: 675925.777,
         });
-        m.expect(input.pipe(distance(Distances.MILES, Distances.METERS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.MILES, Lengths.METERS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -692,7 +692,7 @@ describe('distance', () => {
           b: 176000,
           c: 739200,
         });
-        m.expect(input.pipe(distance(Distances.MILES, Distances.YARDS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.MILES, Lengths.YARDS, 2))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -714,7 +714,7 @@ describe('distance', () => {
           b: 9144.111,
           c: 38405.267,
         });
-        m.expect(input.pipe(distance(Distances.YARDS, Distances.CM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.YARDS, Lengths.CM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -732,7 +732,7 @@ describe('distance', () => {
           b: 300,
           c: 1260,
         });
-        m.expect(input.pipe(distance(Distances.YARDS, Distances.FEET))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.YARDS, Lengths.FEET))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -751,7 +751,7 @@ describe('distance', () => {
           b: 3600,
           c: 15120,
         });
-        m.expect(input.pipe(distance(Distances.YARDS, Distances.INCHES))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.YARDS, Lengths.INCHES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -770,7 +770,7 @@ describe('distance', () => {
           b: 0.091,
           c: 0.384,
         });
-        m.expect(input.pipe(distance(Distances.YARDS, Distances.KM))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.YARDS, Lengths.KM))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -789,7 +789,7 @@ describe('distance', () => {
           b: 91.441,
           c: 384.053,
         });
-        m.expect(input.pipe(distance(Distances.YARDS, Distances.METERS))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.YARDS, Lengths.METERS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -808,7 +808,7 @@ describe('distance', () => {
           b: 0.06,
           c: 0.24,
         });
-        m.expect(input.pipe(distance(Distances.YARDS, Distances.MILES, 2))).toBeObservable(expected);
+        m.expect(input.pipe(length(Lengths.YARDS, Lengths.MILES, 2))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );