+++ libs/rxjs/utility/src/lib/length.ts
@@ -5,16 +5,16 @@
 import { MonoTypeOperatorFunction, Subscribable } from 'rxjs';
 import { createOrReturnObservable } from '../utils/internal';
 import { map, withLatestFrom } from 'rxjs/operators';
-import { Distances } from '../types/distance';
+import { Lengths } from '../types/length';
 import { fromCm, fromFeet, fromInches, fromKm, fromMeters, fromMiles, fromYards } from '../utils/distance';
 
 /**
- * Returns an Observable that converts the source value through [[Distances]] conversion
+ * Returns an Observable that converts the source value through [[Lengths]] conversion
  *
  * @category Conversion
  *
- * @param fromDistance The distance type of the source value
- * @param toDistance The distance type of the output value
+ * @param fromDistance The length type of the source value
+ * @param toDistance The length type of the output value
  * @param precision The number of decimal places to return, default is `3`
  *
  * @example
@@ -22,13 +22,13 @@ import { fromCm, fromFeet, fromInches, fromKm, fromMeters, fromMiles, fromYards 
  * ```ts
  * const source$ = from([100, 200, 300, 500]);
  *
- * source$.pipe(distance(Distances.MILES, Distances.METERS, 0)).subscribe()
+ * source$.pipe(length(Lengths.MILES, Lengths.METERS, 0)).subscribe()
  * ```
  * Output: `160934, 321869, 482803, 804672`
  *
- * @returns Observable that emits a number that is the `from` [[Distances]] converted to the `to` [[Distances]]
+ * @returns Observable that emits a number that is the `from` [[Lengths]] converted to the `to` [[Lengths]]
  */
-export function distance<T extends Distances>(
+export function length<T extends Lengths>(
   fromDistance: Subscribable<T> | T,
   toDistance: Subscribable<T> | T,
   precision: Subscribable<number> | number = 3,
@@ -42,25 +42,25 @@ export function distance<T extends Distances>(
       withLatestFrom(fromDistance$, toDistance$, precision$),
       map<[number, string, string, number], number>(([value, fromDistanceValue, toDistanceValue, precisionValue]) => {
         switch (fromDistanceValue) {
-          case Distances.CM: {
+          case Lengths.CM: {
             return fromCm[toDistanceValue](value, precisionValue);
           }
-          case Distances.FEET: {
+          case Lengths.FEET: {
             return fromFeet[toDistanceValue](value, precisionValue);
           }
-          case Distances.INCHES: {
+          case Lengths.INCHES: {
             return fromInches[toDistanceValue](value, precisionValue);
           }
-          case Distances.KM: {
+          case Lengths.KM: {
             return fromKm[toDistanceValue](value, precisionValue);
           }
-          case Distances.METERS: {
+          case Lengths.METERS: {
             return fromMeters[toDistanceValue](value, precisionValue);
           }
-          case Distances.MILES: {
+          case Lengths.MILES: {
             return fromMiles[toDistanceValue](value, precisionValue);
           }
-          case Distances.YARDS: {
+          case Lengths.YARDS: {
             return fromYards[toDistanceValue](value, precisionValue);
           }
           /* istanbul ignore next-line */